[GB] Add a barebone GameBoy Emulator class
The gbEmulator class will eventually handle all of the Game Boy-family
emulation. For now, it only handles the SGB border option. This allows
for separation of the SGB border option from the internal emulator
setting. This prevents the Game Boy dialog option to display a different
setting from the one set up by the user.

In the long term, having one class to handle all of the emulator data
for a specific system will allow for removing most of the global state
and easily setting up multiple emulator instances.